Increase matrix index size to i32
Now that we're basically always running with --slim, the size of the
matrix index isn't that important it seems.  On a reasonably sized
input it made no noticable difference.

+++ mars-core/src/sneaky_matrix.rs
@@ -2,7 +2,7 @@ use crate::permutation::Permutation;
 use serde::{Deserialize, Serialize};
 
 /// Index type for the matrix.
-pub type CI = i16;
+pub type CI = i32;
 
 /// A sparse column representation for a Boolean matrix.
 ///
